backend/requirements.txt
chromadb==1.4.0
docling==2.66.0
langchain_text_splitters==1.1.0
protobuf==6.33.2
pydantic==2.12.5
pypdf==6.5.0
python-dotenv==1.2.1
Requests==2.32.5
google-genai==0.3.0
fastapi[standard]>=0.115.0

# HuggingFace model download + caching stack (CRITICAL for Docling)
huggingface_hub>=0.23.0
hf_xet
safetensors
transformers

# Native PDF rendering backend for Docling (CRITICAL for speed & tables)
pypdfium2
